requirements.txt
fastapi
uvicorn
jinja2
python-multipart
requests
aiohttp
transformers
torch
sentencepiece
sqlalchemy[asyncio]
aiosqlite
redis
python-dotenv